pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kuberay-sdk"
version = "0.1.0"
description = "A user-friendly Python SDK for managing Ray clusters, jobs, and services on Kubernetes and OpenShift via KubeRay."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "KubeRay SDK Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "kubernetes>=28.1.0",
    "kube-authkit>=0.4.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kuberay_sdk"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B017",   # assertRaises(Exception) in tests is intentional
    "TC001",  # typing-only first-party imports not worth moving
    "TC003",  # typing-only stdlib imports not worth moving
    "RUF043", # pytest.raises ambiguous pattern acceptable in tests
    "SIM105", # contextlib.suppress vs try/except is style preference
    "N818",   # error suffix naming convention not applicable here
]

[tool.ruff.lint.isort]
known-first-party = ["kuberay_sdk"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "kubernetes.*",
    "kube_authkit.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
